+++ src/components/shared/CardGrid/CardGrid.tsx
@@ -5,6 +5,7 @@ import Card from "./Card";
 import Image from "./../../shared/Image";
 import TextFormatted from "../TextFormatted";
 import { ImageType } from "../Image/ImageTypes";
+import { ButtonLinkProps as ButtonLinkType } from "./../ButtonLink/ButtonLink";
 
 export type CardGridLayoutTypes = "row" | "column" | "column_two_featured";
 
@@ -40,6 +41,7 @@ export interface CardItem {
   description: string;
   image?: ImageType;
   link: string;
+  buttonLinks?: ButtonLinkType[];
 }
 
 export default function CardGrid({
@@ -126,7 +128,7 @@ export default function CardGrid({
       >
         {items &&
           items.map((item: CardItem, index) => {
-            const { id, title, description, link, image } = item;
+            const { id, title, description, link, image, buttonLinks } = item;
             return (
               <Box
                 as="li"
@@ -157,6 +159,7 @@ export default function CardGrid({
                       />
                     ),
                   })}
+                  buttonLinks={buttonLinks}
                 />
               </Box>
             );