+++ src/components/shared/ContentComponents/ButtonLinks.tsx
@@ -1,18 +1,7 @@
 import * as React from "react";
-import {
-  Box,
-  Flex,
-  Heading,
-  Link,
-  Icon,
-  IconNames,
-} from "@nypl/design-system-react-components";
-
-type ButtonLinkItem = {
-  id: string;
-  link: { title: string; url: string; uri: string };
-  icon: string;
-};
+import { Box, Flex, Heading } from "@nypl/design-system-react-components";
+import ButtonLink from "./../ButtonLink";
+import { ButtonLinkProps as ButtonLinkItem } from "./../ButtonLink/ButtonLink";
 
 interface ButtonLinksProps {
   id: string;
@@ -20,14 +9,6 @@ interface ButtonLinksProps {
   description?: string;
   items: ButtonLinkItem[];
 }
-
-// Lookup table to match drupal strings to the corresponding DS IconNames
-const IconTable: Record<string, IconNames> = {
-  facebook: "legacySocialFacebook",
-  instagram: "socialInstagram",
-  twitter: "socialTwitter",
-  file_type_doc: "fileTypeDoc",
-};
 
 export default function ButtonLinks({ id, heading, items }: ButtonLinksProps) {
   return (
@@ -58,39 +39,7 @@ export default function ButtonLinks({ id, heading, items }: ButtonLinksProps) {
                 listStyleType="none"
                 w={{ sm: "full", md: "fit-content" }}
               >
-                {/* @TODO once we are updating the DS version,
-                replace custome style with type="buttonSecondary"*/}
-                <Link
-                  id={`link-${id}`}
-                  href={item.link.url}
-                  type="action"
-                  w="full"
-                  py="xs"
-                  px="s"
-                  textDecor="none"
-                  color="ui.link.primary"
-                  bg="transparent"
-                  borderStyle="solid"
-                  borderWidth="1px"
-                  borderColor="ui.link.primary"
-                  _hover={{
-                    bg: "rgb(5, 118, 211,0.05 )",
-                    borderColor: "ui.link.secondary",
-                    color: "ui.link.secondary",
-                    textDecor: "none",
-                    svg: {
-                      color: "ui.link.secondary",
-                    },
-                  }}
-                >
-                  <Icon
-                    name={IconTable[item.icon]}
-                    size="small"
-                    align="left"
-                    color="ui.link.primary"
-                  />
-                  {item.link.title}
-                </Link>
+                <ButtonLink id={item.id} link={item.link} icon={item.icon} />
               </Box>
             ))}
         </Flex>