+++ src/pages/_document.tsx
@@ -1,6 +1,7 @@
 const {
   NEXT_PUBLIC_GTM_TRACKING_ID,
   NEXT_PUBLIC_GA_TRACKING_ID,
+  NEXT_PUBLIC_ADOBE_LAUNCH_URL,
   NEXT_PUBLIC_SERVER_ENV,
 } = process.env;
 import { Html, Head, Main, NextScript } from "next/document";
@@ -15,10 +16,14 @@ export default function Document() {
   // Set PROD or QA version of the remote header embed script.
   // This will generate enviornment aware links in the header logins.
   let nyplHeaderScript =
-    "https://header.nypl.org/dgx-header.min.js?skipNav=main-content&urls=absolute";
+    "https://ds-header.nypl.org/header.min.js?containerId=nypl-header";
+  let nyplFooterScript =
+    "https://ds-header.nypl.org/footer.min.js?containerId=nypl-footer";
   if (NEXT_PUBLIC_SERVER_ENV !== "production") {
     nyplHeaderScript =
-      "https://qa-header.nypl.org/dgx-header.min.js?skipNav=main-content&urls=absolute";
+      "https://qa-ds-header.nypl.org/header.min.js?containerId=nypl-header";
+    nyplFooterScript =
+      "https://qa-ds-header.nypl.org/footer.min.js?containerId=nypl-footer";
   }
 
   return (
@@ -33,26 +38,53 @@ export default function Document() {
           id="google-tag-data-layer"
           dangerouslySetInnerHTML={{
             __html: `
-                    window.dataLayer = window.dataLayer || [];
-                    function gtag(){dataLayer.push(arguments);}
-                    gtag('js', new Date());
-                    gtag('require', '${NEXT_PUBLIC_GTM_TRACKING_ID}');
-                    gtag('config', '${NEXT_PUBLIC_GA_TRACKING_ID}', {
-                      page_path: window.location.pathname,
-                      'groups':'default',
-                      'anonymize_ip':true
-                    });
-                  `,
+              window.dataLayer = window.dataLayer || [];
+              function gtag(){dataLayer.push(arguments);}
+              gtag('js', new Date());
+              gtag('require', '${NEXT_PUBLIC_GTM_TRACKING_ID}');
+              gtag('config', '${NEXT_PUBLIC_GA_TRACKING_ID}', {
+                page_path: window.location.pathname,
+                'groups':'default',
+                'anonymize_ip':true
+              });
+            `,
           }}
           strategy="afterInteractive"
         />
-        {/* NYPL Header */}
-        <Script src={nyplHeaderScript} strategy="beforeInteractive" />
+
+        {/* Adobe Analytics: Initial data layer definition. */}
+        <Script
+          id="adobe-analytics-data-layer"
+          strategy="afterInteractive"
+          dangerouslySetInnerHTML={{
+            __html: `
+              window.adobeDataLayer = [];
+              let pageName = document.title.split("|")[0].trim();
+              if (window.location.pathname === "/") {
+                pageName = "Home"
+              }
+              window.adobeDataLayer.push({
+                page_name: pageName,
+                site_section: null,
+              });
+            `,
+          }}
+        />
+        {/* Adobe Analytics: Add launch tag manager script. */}
+        <Script
+          src={NEXT_PUBLIC_ADOBE_LAUNCH_URL}
+          strategy="afterInteractive"
+        />
       </Head>
       <body>
+        <Script src={nyplHeaderScript} strategy="afterInteractive" />
+        <Script src={nyplFooterScript} strategy="afterInteractive" />
         <Main />
         <NextScript />
-        <Script src="https://assets.nypl.org/js/advocacy.js" />
+        <Script
+          src="https://assets.nypl.org/js/advocacy.js"
+          strategy="afterInteractive"
+        />
       </body>
     </Html>
   );